+++ src/lib/declarationHandler.js
@@ -45,6 +45,7 @@ const CLASSES_OF_TRUTH = [
     'NTP',
     'Provision',
     'VLAN',
+    'DNS_Resolver',
     'Trunk',
     'SelfIp',
     'Route',
@@ -65,7 +66,9 @@ const CLASSES_OF_TRUTH = [
     'System',
     'TrafficControl',
     'HTTPD',
-    'SSHD'
+    'SSHD',
+    'Tunnel',
+    'TrafficGroup'
 ];
 
 /**