+++ test/unit/src/lib/inspectHandlerTests.js
@@ -388,7 +388,34 @@ describe('inspectHandler', () => {
 
     describe('declaration verification', () => {
         let listResponses;
-        let failWhenNoPropertyinResponse;
+        let failWhenNoPropertyInResponse;
+        let missedProperties;
+
+        const addMissedProperty = function (uri, property) {
+            missedProperties[uri] = missedProperties[uri] || {};
+            if (missedProperties[uri][property] !== false) {
+                missedProperties[uri][property] = true;
+            }
+        };
+        const removeMissedProperty = function (uri, property) {
+            missedProperties[uri] = missedProperties[uri] || {};
+            missedProperties[uri][property] = false;
+        };
+        const verifyMissedProperties = function () {
+            const missed = [];
+            Object.keys(missedProperties).forEach((uri) => {
+                const props = missedProperties[uri];
+                Object.keys(props).forEach((prop) => {
+                    if (props[prop]) {
+                        missed.push(`${uri}::${prop}`);
+                    }
+                });
+            });
+
+            if (missed.length > 0) {
+                throw new Error(`These properties have no response data: ${JSON.stringify(missed)}`);
+            }
+        };
 
         const pathsToIgnore = configItems.filter(item => item.declaration === false).map(item => item.path);
         const deviceName = 'device1';
@@ -420,8 +447,12 @@ describe('inspectHandler', () => {
                         ret.name = data.name;
                     }
                     $select.forEach((key) => {
-                        if (key !== 'name' && failWhenNoPropertyinResponse) {
-                            assert.notStrictEqual(data[key], undefined, `Should have '${key}' in response data for '${parsedURL.pathname}' in listResponses`);
+                        if (key !== 'name' && failWhenNoPropertyInResponse) {
+                            if (typeof data[key] === 'undefined') {
+                                addMissedProperty(parsedURL.pathname, key);
+                            } else {
+                                removeMissedProperty(parsedURL.pathname, key);
+                            }
                         }
                         ret[key] = data[key];
                     });
@@ -472,6 +503,43 @@ describe('inspectHandler', () => {
                 nameServers: ['172.27.1.1'],
                 search: ['localhost']
             },
+            '/tm/net/dns-resolver': [
+                {
+                    name: 'testDnsResolver',
+                    answerDefaultZones: 'no',
+                    cacheSize: 5767168,
+                    forwardZones: [
+                        {
+                            name: 'amazonaws.com',
+                            nameservers: [
+                                {
+                                    name: '8.8.8.8:53'
+                                },
+                                {
+                                    name: '8.8.8.7:53'
+                                }
+                            ]
+                        },
+                        {
+                            name: 'idservice.net',
+                            nameservers: [
+                                {
+                                    name: '8.8.4.4:53'
+                                },
+                                {
+                                    name: '8.8.4.3:53'
+                                }
+                            ]
+                        }
+                    ],
+                    randomizeQueryNameCase: 'yes',
+                    routeDomain: 0,
+                    useIpv4: 'yes',
+                    useIpv6: 'yes',
+                    useTcp: 'yes',
+                    useUdp: 'yes'
+                }
+            ],
             '/tm/net/trunk': [
                 {
                     name: 'testTrunk',
@@ -493,7 +561,10 @@ describe('inspectHandler', () => {
                     interfacesReference: {
                         link: 'https://localhost/mgmt/tm/net/vlan/~Common~internalVlan/interfaces'
                     },
-                    cmpHash: 'default'
+                    cmpHash: 'default',
+                    failsafe: 'enabled',
+                    failsafeAction: 'reboot',
+                    failsafeTimeout: 3600
                 },
                 {
                     name: 'externalVlan',
@@ -502,7 +573,10 @@ describe('inspectHandler', () => {
                     interfacesReference: {
                         link: 'https://localhost/mgmt/tm/net/vlan/~Common~externalVlan/interfaces'
                     },
-                    cmpHash: 'src-ip'
+                    cmpHash: 'src-ip',
+                    failsafe: 'disabled',
+                    failsafeAction: 'failover-restart-tm',
+                    failsafeTimeout: 90
                 }
             ],
             '/tm/net/vlan/~Common~externalVlan/interfaces': [
@@ -540,6 +614,12 @@ describe('inspectHandler', () => {
                     name: 'testRoute2',
                     gw: '11.0.0.11',
                     network: '30.0.0.0/24',
+                    mtu: 0
+                },
+                {
+                    name: 'testRoute3',
+                    interface: '/Common/tunnel',
+                    network: '1.2.3.4/32',
                     mtu: 0
                 }
             ],
@@ -804,7 +884,17 @@ describe('inspectHandler', () => {
                 bannerText: 'This is the banner text',
                 inactivityTimeout: 10000,
                 include: 'Ciphers aes128-ctr,aes256-ctr\nLoginGraceTime 10\nMACs hmac-sha1\nMaxAuthTries 5\nMaxStartups 3\nProtocol 1\n'
-            }
+            },
+            '/tm/net/tunnels/tunnel': [
+                {
+                    name: 'tunnel',
+                    mtu: 0,
+                    profile: '/Common/tcp-forward',
+                    tos: 'preserve',
+                    usePmtu: 'enabled',
+                    autoLasthop: 'default'
+                }
+            ]
         });
 
         // PURPOSE: to be sure that all properties (we are expecting) are here
@@ -851,6 +941,33 @@ describe('inspectHandler', () => {
                             search: ['localhost'],
                             class: 'DNS'
                         },
+                        testDnsResolver: {
+                            answerDefaultZones: false,
+                            cacheSize: 5767168,
+                            forwardZones: [
+                                {
+                                    name: 'amazonaws.com',
+                                    nameservers: [
+                                        '8.8.8.8:53',
+                                        '8.8.8.7:53'
+                                    ]
+                                },
+                                {
+                                    name: 'idservice.net',
+                                    nameservers: [
+                                        '8.8.4.4:53',
+                                        '8.8.4.3:53'
+                                    ]
+                                }
+                            ],
+                            randomizeQueryNameCase: true,
+                            routeDomain: 0,
+                            useIpv4: true,
+                            useIpv6: true,
+                            useTcp: true,
+                            useUdp: true,
+                            class: 'DNS_Resolver'
+                        },
                         testTrunk: {
                             distributionHash: 'dst-mac',
                             interfaces: [],
@@ -870,7 +987,10 @@ describe('inspectHandler', () => {
                                 { name: '2.2', tagged: false }
                             ],
                             class: 'VLAN',
-                            cmpHash: 'default'
+                            cmpHash: 'default',
+                            failsafeEnabled: true,
+                            failsafeAction: 'reboot',
+                            failsafeTimeout: 3600
                         },
                         externalVlan: {
                             mtu: 1500,
@@ -880,7 +1000,10 @@ describe('inspectHandler', () => {
                                 { name: '1.2', tagged: false }
                             ],
                             class: 'VLAN',
-                            cmpHash: 'src-ip'
+                            cmpHash: 'src-ip',
+                            failsafeEnabled: false,
+                            failsafeAction: 'failover-restart-tm',
+                            failsafeTimeout: 90
                         },
                         internalSelfIp: {
                             address: '10.0.0.2/24',
@@ -905,6 +1028,12 @@ describe('inspectHandler', () => {
                         testRoute2: {
                             gw: '11.0.0.11',
                             network: '30.0.0.0/24',
+                            mtu: 0,
+                            class: 'Route'
+                        },
+                        testRoute3: {
+                            target: 'tunnel',
+                            network: '1.2.3.4/32',
                             mtu: 0,
                             class: 'Route'
                         },
@@ -1202,6 +1331,14 @@ describe('inspectHandler', () => {
                             maxAuthTries: 5,
                             maxStartups: '3',
                             protocol: 1
+                        },
+                        tunnel: {
+                            class: 'Tunnel',
+                            tunnelType: 'tcp-forward',
+                            mtu: 0,
+                            usePmtu: true,
+                            typeOfService: 'preserve',
+                            autoLastHop: 'default'
                         }
                     }
                 }
@@ -1215,14 +1352,19 @@ describe('inspectHandler', () => {
 
         beforeEach(() => {
             // skip data asserts
+            missedProperties = {};
             expectedDeclaration = undefined;
-            failWhenNoPropertyinResponse = false;
+            failWhenNoPropertyInResponse = false;
             inspectHandler = new InspectHandler();
             listResponses = defaultResponses();
         });
 
         after(() => {
             sinon.restore();
+        });
+
+        afterEach(() => {
+            verifyMissedProperties();
         });
 
         it('should verify that all items from configItems.json are covered', () => {
@@ -1283,7 +1425,7 @@ describe('inspectHandler', () => {
         });
 
         it('should verify declaration from response', () => {
-            failWhenNoPropertyinResponse = true;
+            failWhenNoPropertyInResponse = true;
             return inspectHandler.process()
                 .then((data) => {
                     expectedDeclaration = referenceDeclaration;